move toTcx to createTcxFile
