+++ frontend/src/utils.ts
@@ -1,5 +1,4 @@
-import { lapToTCX } from './createTcxFile';
-import { Lap, Workout } from './types';
+import { Workout } from './types';
 
 export const randomIntFromIntervalBasedOnPrev = (
   min: number,
@@ -29,49 +28,6 @@ export const removeDuplicateWords = (words: string[]) =>
   );
 
 export const mailIsValid = (mail: string) => /.+@.+\..+/.test(mail.trim());
-
-export const toTCX = (laps: Lap[]) => {
-  const startTime = laps[0].dataPoints[0].timeStamp;
-  const output = [
-    `<?xml version="1.0" encoding="UTF-8"?>`,
-    `<TrainingCenterDatabase xmlns="http://www.garmin.com/xmlschemas/TrainingCenterDatabase/v2"> `,
-    `  <Activities>`,
-    `    <Activity Sport="Biking">`,
-    `      <Id>${startTime.toISOString()}</Id>`,
-    laps.map((lap) => lapToTCX(lap)).join('\n'),
-    `    </Activity>`,
-    `  </Activities>`,
-    `  <Author xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="Application_t">`,
-    `    <Name>Dundring</Name>`,
-    `    <Build>`,
-    `      <Version>`,
-    `        <VersionMajor>0</VersionMajor>`,
-    `        <VersionMinor>0</VersionMinor>`,
-    `      </Version>`,
-    `    </Build>`,
-    `    <LangID>EN</LangID>`,
-    `    <PartNumber>XXX-XXXXX-XX</PartNumber>`,
-    `  </Author>`,
-    `</TrainingCenterDatabase>`,
-  ].join('\n');
-
-  const url = window.URL.createObjectURL(new Blob([output]));
-  const link = document.createElement('a');
-
-  const filename = `dundring_${formatDateForFilename(startTime)}`;
-
-  link.href = url;
-  link.setAttribute('download', `${filename}.tcx`);
-
-  // Append to html link element page
-  document.body.appendChild(link);
-
-  // Start download
-  link.click();
-
-  // Clean up and remove the link
-  link.parentNode?.removeChild(link);
-};
 
 export interface HoursMinutesAndSeconds {
   hours: number;
@@ -108,11 +64,4 @@ export const timestampToFormattedHHMMSS = (timestamp: Date) => {
 export const getTotalWorkoutTime = (workout: Workout) =>
   workout.parts.reduce((sum, part) => sum + part.duration, 0);
 
-const padLeadingZero = (nr: number) => (nr < 10 ? '0' + nr : nr);
-
-const formatDateForFilename = (date: Date): string => {
-  const yyyymmdd = date.toISOString().split('T')[0];
-  const hhmm =
-    padLeadingZero(date.getHours()) + '' + padLeadingZero(date.getMinutes());
-  return yyyymmdd + 'T' + hhmm;
-};
+export const padLeadingZero = (nr: number) => (nr < 10 ? '0' + nr : nr);
